Add automatic packages for NCrunch
Adds automatic packages for NCrunch for VS2012, VS2013 and Grid Node
Server.
Uses Package IDs that conform to the naming rules
(https://github.com/chocolatey/choco/wiki/CreatePackages#naming-your-package)
Changes the existing package for NCrunch for VS2015 to be an automatic
package.
References the actual release notes for the version.
Deprecates the old packages with the incorrect package IDs as per
https://github.com/chocolatey/choco/wiki/How-To-Deprecate-A-Chocolatey-Package